Split typechecker resolver lookup helpers
